+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Breaking changes: Update `LinkList` component with styleguide V2.
+- Breaking change: Update `LinkList` component with styleguide V2.
   Replace `@include k-LinkList(( colors: ( ))` by `@include k-LinkList;`.
 - Feature: Add `font2` prop in `Text` component.
 - Feature: Create `HorizontalStroke` component.